+++ fj-doc-freemarker/src/main/java/org/fugerit/java/doc/freemarker/fun/SimpleMessageFun.java
@@ -12,9 +12,7 @@ public class SimpleMessageFun implements TemplateMethodModelEx {
 
 	@Override
 	public Object exec(@SuppressWarnings("rawtypes") List arguments) throws TemplateModelException {
-		if ( arguments.isEmpty() ) {
-			throw new TemplateModelException( "At least one parameter is needed" );
-		}
+		FMFunHelper.checkFirstRequred( arguments );
 		String key = ((TemplateScalarModel)arguments.get( 0 )).getAsString();
 		Object[] args = new Object[ arguments.size()-1 ];
 		for ( int k=1; k<arguments.size(); k++ ) {