+++ fj-doc-freemarker/src/main/java/org/fugerit/java/doc/freemarker/fun/TextWrapFun.java
@@ -1,0 +1,21 @@
+package org.fugerit.java.doc.freemarker.fun;
+
+import freemarker.template.SimpleScalar;
+import freemarker.template.TemplateMethodModelEx;
+import freemarker.template.TemplateModelException;
+import freemarker.template.TemplateScalarModel;
+import org.fugerit.java.doc.base.helper.TextWrapHelper;
+
+import java.text.MessageFormat;
+import java.util.List;
+
+public class TextWrapFun implements TemplateMethodModelEx {
+
+	@Override
+	public Object exec(@SuppressWarnings("rawtypes") List arguments) throws TemplateModelException {
+		FMFunHelper.checkFirstRequred( arguments );
+		String input = ((TemplateScalarModel)arguments.get( 0 )).getAsString();
+		return new SimpleScalar(TextWrapHelper.padZeroWithSpace( input ));
+	}
+	
+}